Separate traits from impl with feature flag
