+++ p2panda-group/Cargo.toml
@@ -15,14 +15,26 @@ all-features = true
 [lints]
 workspace = true
 
+[features]
+default = ["provider"]
+provider = [
+  "dep:chacha20poly1305",
+  "dep:curve25519-dalek",
+  "dep:libcrux",
+  "dep:libcrux-ed25519",
+  "dep:libcrux-traits",
+  "dep:rand_chacha",
+  "dep:subtle",
+]
+
 [dependencies]
-chacha20poly1305 = "0.10.1"
-curve25519-dalek = "4.1.3"
-libcrux = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888" }
-libcrux-ed25519 = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888" }
-libcrux-traits = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888" }
-rand_chacha = { version = "0.9.0", features = ["os_rng"] }
+chacha20poly1305 = { version = "0.10.1", optional = true }
+curve25519-dalek = { version = "4.1.3", optional = true }
+libcrux = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888", optional = true }
+libcrux-ed25519 = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888", optional = true }
+libcrux-traits = { git = "https://github.com/cryspen/libcrux", rev = "cf292a2829172cf9b36390f2f05f7a32b2304888", optional = true }
+rand_chacha = { version = "0.9.0", features = ["os_rng"], optional = true }
 serde = { version = "1.0.219", features = ["derive"] }
-subtle = "2.6.1"
+subtle = { version = "2.6.1", optional = true }
 thiserror = "2.0.12"
 zeroize = { version = "1.8.1", features = ["derive"] }
