+++ p2panda-group/src/crypto/ed25519.rs
@@ -94,7 +94,7 @@ pub enum SignatureError {
 
 #[cfg(test)]
 mod tests {
-    use crate::crypto::{Provider, RandProvider};
+    use crate::provider::{Provider, RandProvider};
 
     use super::{SignatureError, SigningKey};
