+++ p2panda-group/src/lib.rs
@@ -1,3 +1,5 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+#[cfg(feature = "provider")]
 pub mod crypto;
+pub mod provider;
